flask-template/requirements.txt
flask
SQLAlchemy
pymysql
gtts
tqdm
python-dotenv
